Factor out bootstrap into a separate class
